// src/types/frondesk/appointment/types.ts
export enum EnumAppointmentRepeatType {
  DoesNotRepeat = 'Does not repeat',
  Daily = 'Daily',
  Weekly = 'Weekly',
  Weekends = 'Weekends',
  Monthly = 'Monthly',
  Weekdays = 'Weekdays',
  Annually = 'Annually',
  Custom = 'Custom',
}

export enum EnumAppointmentType {
  WalkIn = 'Walk-in',
  Referral = 'Referral',
  Consultation = 'Consultation',
  FollowUp = 'Follow-up',
  MedicalExam = 'Medical-exam',
}